+++ frontend/src/views/ProducerFormPage/SummaryTab.vue
